cicd/requirements.txt
# AWS CDK dependencies
aws-cdk-lib>=2.100.0
constructs>=10.0.0

# Thomson Reuters CDK library
tr-cdk-lib>=1.0.0

# Python dependencies
boto3>=1.26.0
